+++ src/resources/views/edit.blade.php
@@ -26,10 +26,10 @@
 		    <div class="box-header with-border">
 		      <h3 class="box-title">{{ trans('backpack::crud.edit') }}</h3>
 		    </div>
-		    <div class="box-body">
+		    <div class="box-body row">
 		      <!-- load the view from the application if it exists, otherwise load the one in the package -->
-		      @if(view()->exists('vendor.dick.crud.form_content'))
-		      	@include('vendor.dick.crud.form_content')
+		      @if(view()->exists('vendor.backpack.crud.form_content'))
+		      	@include('vendor.backpack.crud.form_content')
 		      @else
 		      	@include('crud::form_content', ['fields' => $crud->getFields('update', $entry->getKey())])
 		      @endif