+++ src/resources/views/fields/date.blade.php
@@ -1,18 +1,15 @@
 <!-- html5 date input -->
-  <div class="form-group">
+<div @include('crud::inc.fieldWrapperAttributes') >
     <label>{{ $field['label'] }}</label>
     <input
         type="date"
-        class="form-control"
-
-        @foreach ($field as $attribute => $value)
-            @if (is_string($attribute) && is_string($value))
-                @if ($attribute=='value')
-                    value="{{ old($field['name']) ? old($field['name']) : $value }}"
-                @else
-                    {{ $attribute }}="{{ $value }}"
-                @endif
-            @endif
-        @endforeach
+        name="{{ $field['name'] }}"
+        value="{{ old($field['name']) ? old($field['name']) : (isset($field['value']) ? $field['value'] : (isset($field['default']) ? $field['default'] : '' )) }}"
+        @include('crud::inc.fieldAttributes')
         >
-  </div>
+
+    {{-- HINT --}}
+    @if (isset($field['hint']))
+        <p class="help-block">{!! $field['hint'] !!}</p>
+    @endif
+</div>