+++ src/resources/views/fields/enum.blade.php
@@ -1,15 +1,10 @@
 <!-- enum -->
-  <div class="form-group">
+<div @include('crud::inc.fieldWrapperAttributes') >
     <label>{{ $field['label'] }}</label>
     <?php $entity_model = $crud->model; ?>
     <select
-    	class="form-control"
-
-    	@foreach ($field as $attribute => $value)
-            @if (is_string($attribute) && is_string($value))
-        		{{ $attribute }}="{{ $value }}"
-            @endif
-    	@endforeach
+        name="{{ $field['name'] }}"
+        @include('crud::inc.fieldAttributes')
     	>
 
         @if ($entity_model::isColumnNullable($field['name']))
@@ -26,4 +21,9 @@
 	    		@endforeach
 	    	@endif
 	</select>
-  </div>
+
+    {{-- HINT --}}
+    @if (isset($field['hint']))
+        <p class="help-block">{!! $field['hint'] !!}</p>
+    @endif
+</div>