+++ src/resources/views/fields/select_from_array.blade.php
@@ -1,14 +1,9 @@
 <!-- select -->
-  <div class="form-group">
+<div @include('crud::inc.fieldWrapperAttributes') >
     <label>{{ $field['label'] }}</label>
     <select
-    	class="form-control"
-
-    	@foreach ($field as $attribute => $value)
-            @if (is_string($value))
-    		{{ $attribute }}="{{ $value }}"
-            @endif
-    	@endforeach
+        name="{{ $field['name'] }}"
+        @include('crud::inc.fieldAttributes')
     	>
 
         @if (isset($field['allows_null']) && $field['allows_null']==true)
@@ -25,5 +20,9 @@
 	    		@endforeach
 	    	@endif
 	</select>
-  </div>
-  
+
+    {{-- HINT --}}
+    @if (isset($field['hint']))
+        <p class="help-block">{!! $field['hint'] !!}</p>
+    @endif
+</div>