+++ src/resources/views/fields/select_multiple.blade.php
@@ -1,28 +1,28 @@
 <!-- select multiple -->
-  <div class="form-group">
+<div @include('crud::inc.fieldWrapperAttributes') >
     <label>{{ $field['label'] }}</label>
     <select
     	class="form-control"
-
-    	@foreach ($field as $attribute => $value)
-            @if (is_string($attribute))
-        		@if ($attribute=='name')
-        			{{ $attribute }}="{{ $value }}[]"
-        		@else
-        			{{ $attribute }}="{{ $value }}"
-        		@endif
-            @endif
-    	@endforeach
+        name="{{ $field['name'] }}"
+        @include('crud::inc.fieldAttributes')
     	multiple>
+
     	<option value="">-</option>
-	    	@if (isset($field['model']))
-	    		@foreach ($field['model']::all() as $connected_entity_entry)
-	    			<option value="{{ $connected_entity_entry->getKey() }}"
-						@if ( (isset($field['value']) && in_array($connected_entity_entry->getKey(), $field['value']->lists('id', 'id')->toArray())) || ( old( $field["name"] ) && in_array($connected_entity_entry->getKey(), old( $field["name"])) ) )
-							 selected
-						@endif
-	    			>{{ $connected_entity_entry->{$field['attribute']} }}</option>
-	    		@endforeach
-	    	@endif
+
+    	@if (isset($field['model']))
+    		@foreach ($field['model']::all() as $connected_entity_entry)
+    			<option value="{{ $connected_entity_entry->getKey() }}"
+					@if ( (isset($field['value']) && in_array($connected_entity_entry->getKey(), $field['value']->lists('id', 'id')->toArray())) || ( old( $field["name"] ) && in_array($connected_entity_entry->getKey(), old( $field["name"])) ) )
+						 selected
+					@endif
+    			>{{ $connected_entity_entry->{$field['attribute']} }}</option>
+    		@endforeach
+    	@endif
+
 	</select>
-  </div>
+
+    {{-- HINT --}}
+    @if (isset($field['hint']))
+        <p class="help-block">{!! $field['hint'] !!}</p>
+    @endif
+</div>