+++ src/resources/views/fields/text.blade.php
@@ -1,21 +1,34 @@
 <!-- text input -->
-  <div class="form-group">
-    <label>{{ $field['label'] }}</label>
+<div @include('crud::inc.fieldWrapperAttributes') >
+    <label>{!! $field['label'] !!}</label>
     <input
         type="text"
-        class="form-control"
+        name="{{ $field['name'] }}"
+        value="{{ old($field['name']) ? old($field['name']) : (isset($field['value']) ? $field['value'] : (isset($field['default']) ? $field['default'] : '' )) }}"
+        @include('crud::inc.fieldAttributes')
+    >
 
-        @foreach ($field as $attribute => $value)
-            @if (is_string($attribute) && is_string($value))
-               @if($attribute != 'value' && $attribute != 'type')
-                    {{ $attribute }}="{{ $value }}"
-                @endif
-            @endif
-        @endforeach
-
-        value="{{ old($field['name']) ? old($field['name']) : (isset($field['value'])?$field['value']:'') }}"
-        >
+    {{-- HINT --}}
     @if (isset($field['hint']))
         <p class="help-block">{!! $field['hint'] !!}</p>
     @endif
-  </div>
+</div>
+
+
+{{-- FIELD EXTRA CSS  --}}
+{{-- push things in the after_styles section --}}
+
+    {{-- @push('crud_fields_styles')
+        <!-- no styles -->
+    @endpush --}}
+
+
+{{-- FIELD EXTRA JS --}}
+{{-- push things in the after_scripts section --}}
+
+    {{-- @push('crud_fields_scripts')
+        <!-- no scripts -->
+    @endpush --}}
+
+
+{{-- Note: you can use @if ($crud->checkIfFieldIsFirstOfItsType($field, $fields)) to only load some CSS/JS once, even though there are multiple instances of it --}}